Refactor ViewSingleRollCallHeaderRight
